+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # 2.x branch
 ## 2.31 branch
+### 2.31.7
+* new passwords also require the presence of a capital letter (in addition to the previous rules);
+* added the `AuthenticationComponent`, with the `afterIdentify()` event method. This event is dispatched after the
+  identity and takes care of its management. Some code has been moved from the `UsersController::login()` method;
+* custom messages have been restored when, at login, the user account has been banned or is not yet active. Improved all
+  tests related to the login code;
+* improved admin view for users;
+* every user can view his last logins;
+* vertical alignment is used for admin interface tables.
+
 ### 2.31.6
 * fixed the `connector.minimal.php` file for ElFinder and the `FixElFinderCommand`: the Composer autoload is now
   correctly loaded via absolute path;